# Doc/OpenAPI/help/openapi.yaml
{
  "host": "",
  "basePath": "",
  "schemes": [
    "http"
  ],
  "openapi": "3.0.1",
  "info": {
    "title": "B8AIadmin 接口文档",
    "description": "基于 SaiAdmin/Webman 控制器注解生成，供接口调试与 uniapp 自动接口生成使用。",
    "version": "1.0.7"
  },
...
  "paths": {
    "/app/help/chat/overview": {
      "get": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "聊天模式概览",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "modes": {
                      "type": "array",
                      "description": "启用中的互动角色、能力开关、标签、机器人形象及最近会话",
                      "items": {
                        "type": "string"
                      }
                    },
                    "recent_sessions": {
                      "type": "array",
                      "description": "最近会话",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/config": {
      "get": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "聊天配置",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          },
          {
            "name": "chat_mode",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "互动角色编码，对应 sa_ai_persona.code"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "type": "array",
                      "description": "会员自定义聊天提示词配置",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      },
      "post": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "保存聊天配置",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "chat_mode": {
                    "type": "string",
                    "description": "互动角色编码，对应 sa_ai_persona.code"
                  },
                  "prompt_text": {
                    "type": "string",
                    "description": "用户自定义提示词"
                  }
                },
                "required": [
                  "chat_mode",
                  "prompt_text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "配置ID",
                      "format": "int32"
                    },
                    "chat_mode": {
                      "type": "string",
                      "description": "聊天模式"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/sessions": {
      "get": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "聊天会话列表",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          },
          {
            "name": "chat_mode",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "互动角色编码，对应 sa_ai_persona.code"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "关键词"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "页码",
            "format": "int32"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "每页数量",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "type": "array",
                      "description": "会话列表",
                      "items": {
                        "type": "string"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "总数",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/session": {
      "post": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "创建或更新聊天会话",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "会话ID，传入时更新",
                    "format": "int32"
                  },
                  "chat_mode": {
                    "type": "string",
                    "description": "互动角色编码，对应 sa_ai_persona.code"
                  },
                  "session_name": {
                    "type": "string",
                    "description": "会话名称"
                  },
                  "is_pinned": {
                    "type": "integer",
                    "description": "是否置顶 1是 2否",
                    "format": "int32"
                  }
                },
                "required": [
                  "chat_mode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "会话ID",
                      "format": "int32"
                    },
                    "session_name": {
                      "type": "string",
                      "description": "会话名称"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/session/delete": {
      "post": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "删除聊天会话",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "会话ID",
                    "format": "int32"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "string",
                      "description": "是否删除"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/records": {
      "get": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "聊天记录列表",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "会话ID",
            "format": "int32"
          },
          {
            "name": "chat_mode",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "互动角色编码，对应 sa_ai_persona.code"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "页码",
            "format": "int32"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "每页数量",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "type": "array",
                      "description": "聊天记录列表",
                      "items": {
                        "type": "string"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "总数",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/record": {
      "post": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "保存用户聊天消息",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "session_id": {
                    "type": "integer",
                    "description": "会话ID",
                    "format": "int32"
                  },
                  "content": {
                    "type": "string",
                    "description": "消息内容"
                  },
                  "content_type": {
                    "type": "string",
                    "description": "内容类型 text/image/file/voice"
                  }
                },
                "required": [
                  "session_id",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "消息ID",
                      "format": "int32"
                    },
                    "message_time": {
                      "type": "string",
                      "description": "消息时间",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
    "/app/help/chat/send": {
      "post": {
        "tags": [
          "helpsupport-api.AI聊天.ChatController"
        ],
        "summary": "发送在线AI消息",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "身份令牌Token"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "session_id": {
                    "type": "integer",
                    "description": "会话ID，不传时按 chat_mode 创建新会话",
                    "format": "int32"
                  },
                  "chat_mode": {
                    "type": "string",
                    "description": "互动角色编码，创建新会话时必填"
                  },
                  "content": {
                    "type": "string",
                    "description": "用户消息内容"
                  },
                  "config_id": {
                    "type": "integer",
                    "description": "可选 SaiAI 模型配置ID",
                    "format": "int32"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "type": "object",
                      "description": "会话信息"
                    },
                    "user_record": {
                      "type": "object",
                      "description": "用户消息"
                    },
                    "assistant_record": {
                      "type": "object",
                      "description": "AI 回复消息"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录令牌无效"
          }
        }
      }
    },
...
  },
...
}